requirements.txt
flask
torch
torchvision
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
opencv-python
Pillow
numpy